# docs/openapi_spec.yaml
openapi: 3.1.0
info:
  title: Trendit API
  description: Comprehensive Reddit Data Collection and Analysis Platform
  version: 1.0.0
paths:
...
  /api/collect/jobs:
    post:
      tags:
      - collection
      summary: Create Collection Job
      description: 'Create a new persistent collection job


        This endpoint creates a collection job that will run in the background

        and store all collected data in the database for later analysis.'
      operationId: create_collection_job_api_collect_jobs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionJobRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionJobResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - collection
      summary: List Collection Jobs
      description: List collection jobs with optional filtering
      operationId: list_collection_jobs_api_collect_jobs_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/JobStatus'
          - type: 'null'
          title: Status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionJobListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CollectionJobListResponse:
      properties:
        jobs:
          items:
            $ref: '#/components/schemas/CollectionJobResponse'
          type: array
          title: Jobs
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
      type: object
      required:
      - jobs
      - total
      - page
      - per_page
      title: CollectionJobListResponse
      description: Response for listing collection jobs
    CollectionJobRequest:
      properties:
        subreddits:
          items:
            type: string
          type: array
          title: Subreddits
          description: List of subreddit names
        sort_types:
          items:
            $ref: '#/components/schemas/SortType'
          type: array
          title: Sort Types
          description: Sort types to use
          default:
          - hot
        time_filters:
          items:
            $ref: '#/components/schemas/TimeFilter'
          type: array
          title: Time Filters
          description: Time filters to apply
          default:
          - week
        post_limit:
          type: integer
          maximum: 10000.0
          minimum: 1.0
          title: Post Limit
          description: Maximum posts to collect
          default: 100
        comment_limit:
          type: integer
          maximum: 1000.0
          minimum: 0.0
          title: Comment Limit
          description: Maximum comments per post
          default: 50
        max_comment_depth:
          type: integer
          maximum: 10.0
          minimum: 0.0
          title: Max Comment Depth
          description: Maximum comment thread depth
          default: 3
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
          description: Search keywords
        min_score:
          type: integer
          title: Min Score
          description: Minimum post score
          default: 0
        min_upvote_ratio:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Min Upvote Ratio
          description: Minimum upvote ratio
          default: 0.0
        date_from:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
          description: Start date filter
        date_to:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
          description: End date filter
        exclude_nsfw:
          type: boolean
          title: Exclude Nsfw
          description: Exclude NSFW content
          default: true
        anonymize_users:
          type: boolean
          title: Anonymize Users
          description: Anonymize user information
          default: true
      type: object
      required:
      - subreddits
      title: CollectionJobRequest
      description: Request model for creating a new collection job
    CollectionJobResponse:
      properties:
        id:
          type: integer
          title: Id
        job_id:
          type: string
          title: Job Id
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: integer
          title: Progress
        total_expected:
          type: integer
          title: Total Expected
        collected_posts:
          type: integer
          title: Collected Posts
        collected_comments:
          type: integer
          title: Collected Comments
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        subreddits:
          items:
            type: string
          type: array
          title: Subreddits
        post_limit:
          type: integer
          title: Post Limit
      type: object
      required:
      - id
      - job_id
      - status
      - progress
      - total_expected
      - collected_posts
      - collected_comments
      - error_message
      - created_at
      - started_at
      - completed_at
      - subreddits
      - post_limit
      title: CollectionJobResponse
      description: Response model for collection job information
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobStatus:
      type: string
      enum:
      - pending
      - running
      - completed
      - failed
      - cancelled
      title: JobStatus
...
    SortType:
      type: string
      enum:
      - hot
      - new
      - top
      - rising
      - controversial
      title: SortType
    TimeFilter:
      type: string
      enum:
      - hour
      - day
      - week
      - month
      - year
      - all
      title: TimeFilter
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...